optimize: 内置权限管理下线 #3338
