+++ src/apimachinery/apiserver/api.go
@@ -53,78 +53,6 @@ func (a *apiServer) SearchDefaultApp(ctx context.Context, h http.Header, ownerID
 	return
 }
 
-func (a *apiServer) GetRolePrivilege(ctx context.Context, h http.Header, ownerID, objID, role string) (resp *metadata.RolePriResult, err error) {
-	resp = new(metadata.RolePriResult)
-	subPath := fmt.Sprintf("/topo/privilege/%s/%s/%s", ownerID, objID, role)
-
-	err = a.client.Get().
-		WithContext(ctx).
-		Body(nil).
-		SubResource(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (a *apiServer) GetAppRole(ctx context.Context, h http.Header, params mapstr.MapStr) (resp *metadata.RoleAppResult, err error) {
-
-	resp = new(metadata.RoleAppResult)
-	subPath := fmt.Sprintf("/object/attr/search")
-
-	err = a.client.Post().
-		WithContext(ctx).
-		Body(params).
-		SubResource(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (a *apiServer) GetUserPrivilegeApp(ctx context.Context, h http.Header, ownerID, userName string, params mapstr.MapStr) (resp *metadata.AppQueryResult, err error) {
-
-	resp = new(metadata.AppQueryResult)
-	subPath := fmt.Sprintf("biz/search/%s", ownerID)
-
-	err = a.client.Post().
-		WithContext(ctx).
-		Body(params).
-		SubResource(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (a *apiServer) GetUserPrivilegeConfig(ctx context.Context, h http.Header, ownerID, userName string) (resp *metadata.UserPriviResult, err error) {
-	resp = new(metadata.UserPriviResult)
-	subPath := fmt.Sprintf("topo/privilege/user/detail/%s/%s", ownerID, userName)
-
-	err = a.client.Get().
-		WithContext(ctx).
-		Body(nil).
-		SubResource(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (a *apiServer) GetAllMainLineObject(ctx context.Context, h http.Header, ownerID, userName string) (resp *metadata.MainLineResult, err error) {
-	resp = new(metadata.MainLineResult)
-	subPath := fmt.Sprintf("topo/model/%s", ownerID)
-
-	err = a.client.Get().
-		WithContext(ctx).
-		Body(nil).
-		SubResource(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
 func (a *apiServer) GetObjectData(ctx context.Context, h http.Header, params mapstr.MapStr) (resp *metadata.ObjectAttrBatchResult, err error) {
 	resp = new(metadata.ObjectAttrBatchResult)
 	subPath := fmt.Sprintf("object/search/batch")
