+++ src/auth/parser/topology.go
@@ -46,7 +46,6 @@ func (ps *parseStream) topology() *parseStream {
 		objectUnique().
 		audit().
 		instanceAudit().
-		privilege().
 		fullTextSearch()
 
 	return ps
@@ -2045,30 +2044,6 @@ func (ps *parseStream) instanceAudit() *parseStream {
 }
 
 var (
-	findPrivilege = regexp.MustCompile(`^/api/v3/topo/privilege/.*$`)
-	postPrivilege = regexp.MustCompile(`^/api/v3/topo/privilege/.*$`)
-)
-
-func (ps *parseStream) privilege() *parseStream {
-	if ps.shouldReturn() {
-		return ps
-	}
-
-	if ps.hitRegexp(findPrivilege, http.MethodGet) || ps.hitRegexp(findPrivilege, http.MethodPost) {
-		ps.Attribute.Resources = []meta.ResourceAttribute{
-			{
-				Basic: meta.Basic{
-					Action: meta.SkipAction,
-				},
-			},
-		}
-		return ps
-	}
-
-	return ps
-}
-
-var (
 	fullTextSearchPattern = "/api/v3/find/full_text"
 )
 
@@ -2077,7 +2052,7 @@ func (ps *parseStream) fullTextSearch() *parseStream {
 		return ps
 	}
 
-	if ps.hitRegexp(findPrivilege, http.MethodGet) || ps.hitPattern(fullTextSearchPattern, http.MethodPost) {
+	if ps.hitPattern(fullTextSearchPattern, http.MethodPost) {
 		ps.Attribute.Resources = []meta.ResourceAttribute{
 			{
 				Basic: meta.Basic{
