+++ src/scene_server/admin_server/upgrader/v3.0.8/createtable.go
@@ -134,13 +134,10 @@ var tables = map[string][]dal.Index{
 		dal.Index{Name: "", Keys: map[string]int32{"bk_obj_id": 1, "bk_inst_id": 1}, Background: true},
 	},
 
-	"cc_Privilege":          []dal.Index{},
 	"cc_History":            []dal.Index{},
 	"cc_HostFavourite":      []dal.Index{},
 	"cc_UserAPI":            []dal.Index{},
 	"cc_UserCustom":         []dal.Index{},
-	"cc_UserGroup":          []dal.Index{},
-	"cc_UserGroupPrivilege": []dal.Index{},
 	"cc_idgenerator":        []dal.Index{},
 	"cc_System":             []dal.Index{},
 }
