--- src/web_server/logics/privilege.go
@@ -1,49 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making 蓝鲸 available.
- * Copyright (C) 2017-2018 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * http://opensource.org/licenses/MIT
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package logics
-
-import (
-	"context"
-
-	"configcenter/src/common"
-	"configcenter/src/common/blog"
-	"configcenter/src/common/util"
-	"configcenter/src/web_server/middleware/privilege"
-)
-
-// GetUserAppPri get user privilege
-func (lgc *Logics) GetUserAppPri(ctx context.Context, userName string, ownerID, lang string) (userPriveApp map[int64][]string, rolePrivi map[string][]string, modelConfigPrivi map[string][]string, sysPrivi []string, mainLineObjIDArr []string) {
-	rid := util.ExtractRequestIDFromContext(ctx)
-	p := privilege.NewPrivilege(userName, ownerID, lang)
-	p.Engine = lgc.Engine
-	appRole := p.GetAppRole()
-	rolePrivi = make(map[string][]string)
-	blog.V(5).Infof("get app role result:%v, rid: %s", appRole, rid)
-	if nil != appRole {
-		userPriveApp = p.GetUserPrivilegeApp(appRole)
-		blog.V(5).Infof("get user privi app:%v, rid: %s", userPriveApp, rid)
-		if nil != userPriveApp {
-			for _, role := range appRole {
-				rolePrivilege := p.GetRolePrivilege(common.BKInnerObjIDApp, role)
-				rolePrivi[role] = rolePrivilege
-				blog.V(5).Infof("get role privilege:%v, rid: %s", rolePrivilege, rid)
-			}
-		}
-
-	}
-
-	modelConfigPrivi, sysPrivi = p.GetUserPrivilegeConfig()
-	mainLineObjIDArr = p.GetAllMainLineObject()
-	return
-
-}
