+++ src/web_server/service/index.go
@@ -13,15 +13,8 @@
 package service
 
 import (
-	"encoding/json"
-	"strconv"
-	"strings"
-
 	"configcenter/src/common"
-	"configcenter/src/common/blog"
-	"configcenter/src/common/util"
 	"configcenter/src/common/version"
-	webCommon "configcenter/src/web_server/common"
 
 	"github.com/gin-gonic/gin"
 	"github.com/holmeswang/contrib/sessions"
@@ -29,72 +22,9 @@ import (
 
 // Index html file
 func (s *Service) Index(c *gin.Context) {
-	rid := util.GetHTTPCCRequestID(c.Request.Header)
-	ctx := util.NewContextFromGinContext(c)
 	session := sessions.Default(c)
 	role := session.Get("role")
 	userName, _ := session.Get(common.WEBSessionUinKey).(string)
-	language := webCommon.GetLanguageByHTTPRequest(c)
-
-	if s.Config.Site.AuthScheme == "internal" {
-		userPriviApp, rolePrivilege, modelPrivi, sysPrivi, mainLineObjIDArr := s.Logics.GetUserAppPri(ctx, userName, common.BKDefaultOwnerID, language)
-
-		var strUserPriveApp, strRolePrivilege, strModelPrivi, strSysPrivi, mainLineObjIDStr string
-		if nil == userPriviApp {
-			strUserPriveApp = ""
-		} else {
-			cstrUserPriveApp, _ := json.Marshal(userPriviApp)
-			strUserPriveApp = string(cstrUserPriveApp)
-		}
-
-		if nil == rolePrivilege {
-			strRolePrivilege = ""
-		} else {
-			cstrRolePrivilege, _ := json.Marshal(rolePrivilege)
-			strRolePrivilege = string(cstrRolePrivilege)
-		}
-		if nil == modelPrivi {
-			strModelPrivi = ""
-		} else {
-			cstrModelPrivi, err := json.Marshal(modelPrivi)
-			if err != nil {
-				blog.Errorf("marshal model privilege failed, model privilege: %+v, err: %v, rid: %s", modelPrivi, err, rid)
-			}
-			strModelPrivi = string(cstrModelPrivi)
-		}
-		if nil == sysPrivi {
-			strSysPrivi = ""
-		} else {
-			cstrSysPrivi, err := json.Marshal(sysPrivi)
-			if err != nil {
-				blog.Errorf("marshal system privilege failed, info: %+v, err: %v, rid: %s", sysPrivi, err, rid)
-			}
-			strSysPrivi = string(cstrSysPrivi)
-		}
-
-		mainLineObjIDB, err := json.Marshal(mainLineObjIDArr)
-		if err != nil {
-			blog.Errorf("marshal mainline failed, info: %+v, err: %v, rid: %s", mainLineObjIDArr, err, rid)
-		}
-		mainLineObjIDStr = string(mainLineObjIDB)
-
-		session.Set("userPriviApp", string(strUserPriveApp))
-		session.Set("rolePrivilege", string(strRolePrivilege))
-		session.Set("modelPrivi", string(strModelPrivi))
-		session.Set("sysPrivi", string(strSysPrivi))
-		session.Set("mainLineObjID", string(mainLineObjIDStr))
-		if err := session.Save(); err != nil {
-			blog.Errorf("save session failed, err: %+v, rid: %s", err, rid)
-		}
-
-		// set cookie
-		appIDArr := make([]string, 0)
-		for key := range userPriviApp {
-			appIDArr = append(appIDArr, strconv.FormatInt(key, 10))
-		}
-		appIDStr := strings.Join(appIDArr, "-")
-		c.SetCookie("bk_privi_biz_id", appIDStr, 24*60*60, "", "", false, false)
-	}
 
 	c.HTML(200, "index.html", gin.H{
 		"site":           s.Config.Site.DomainUrl,
